requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.13
langchain-community==0.2.12
langchain-core==0.2.22
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
langsmith==0.1.99
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
